use column_iterator directly, instead segment iterator, and change default value procedure.
